requirements.txt
# Discord
discord.py>=2.3.2
py-cord>=2.4.1

# Database
SQLAlchemy>=2.0.0
alembic>=1.12.0

# Google API
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Configuration
python-dotenv>=1.0.0

# Scheduling
APScheduler>=3.10.4

# Utilities
pytz>=2023.3
python-dateutil>=2.8.2
